Pin the variance floor to float32 explicitly
np.where against a Python float relies on NEP 50 weak-scalar promotion to keep
the result in float32. It does today, and the model arrays are float32
everywhere else on purpose -- half the memory traffic in motion compensation is
gathers over them, and a silent widening to float64 would cost speed with no
signal that anything had changed. Stating the dtype makes that a property of
the code rather than of the numpy version.

No behaviour change: scored identically over a 1500-frame prefix against the
previous commit, 0.34 boxes/frame either way.

Worth recording an environment flake seen while confirming this: one full-video
run returned zero boxes on all 5361 frames at an otherwise normal 20.5ms/frame.
Not a code fault -- the same binary rerun immediately scored 0.773 -- but the
signature is worth knowing. Normal frame time with an empty mask means the
approach ran on frames that decoded as uniform, i.e. get_nvds_buf_surface
handed back a surface it had not filled. run_motion.py swallows a None surface
silently by design, so this failure looks like a bad score rather than an
error. Anything read from that path deserves a second run before it is
believed.

+++ src/deepstream_yolo/approaches/fastmcd.py
@@ -355,7 +355,9 @@ class Approach:
         spread = _block_max((gray - sel_mean) ** 2, b)
         new_var = (var * gain + (1.0 - gain) * spread).astype(np.float32)
         fresh = chosen & (age == 0.0)
-        new_var = np.where(fresh & (new_var < self.init_var), self.init_var, new_var)
+        new_var = np.where(
+            fresh & (new_var < self.init_var), np.float32(self.init_var), new_var
+        ).astype(np.float32)
         np.maximum(new_var, self.min_var, out=new_var)
 
         new_age = age.copy()
